requirements.txt
Flask~=3.1.0
selenium~=4.22.0
beautifulsoup4~=4.12.3
requests~=2.32.3
python-dotenv~=1.0.1